Smal refactor and fixes
