+++ Frontend/src/app/customers/customer-detail/customer-detail.component.html
@@ -36,7 +36,7 @@
         <app-proposition-list [customer]="customer"></app-proposition-list>
       </tab>
       <tab tabTitle="Slutseddel" id="Slutseddel" active="">
-        <app-receipt-list></app-receipt-list>
+        <app-receipt-list [customer]="customer"></app-receipt-list>
       </tab>
       <tab tabTitle="Maskinliste" id="Maskinliste" active="{{equipTab}}">
         <app-equipment-list [customer]="customer"></app-equipment-list>
