+++ Frontend/src/app/receipts/shared/receipt.service.ts
@@ -10,13 +10,9 @@ const url = environment.ApiEndPoint + '/receipts/';
 @Injectable()
 export class ReceiptService {
 
-  private currentReceipt: Receipt;
-  private currentCust: Customer;
 
 
   constructor(private http: HttpClient) {
-    this.currentReceipt = null;
-    this.currentCust = null;
   }
 
   getReceiptsByCustomerId(id: number): Observable<Receipt[]> {
@@ -27,25 +23,6 @@ export class ReceiptService {
     return this.http.post<Receipt>(url, receipt);
   }
 
-  getCreationDateAsEUString(date: Date) {
-    const newDate = new Date(date);
-    let dateString;
-    const options = {year: 'numeric', month: 'numeric', day: 'numeric'};
-    dateString = newDate.toLocaleString('en-GB', options);
-    return dateString;
-  }
-
-  setCurrentReceipt(receipt: Receipt) {
-    this.currentReceipt = receipt;
-  }
-
-  setCurrentCustomer(customer: Customer) {
-    this.currentCust = customer;
-  }
-
-  getCurrentCustomer() {
-    return this.currentCust;
-  }
 
   deleteReceipt(id: number) {
     return this.http.delete<Receipt>(url + id);
@@ -54,13 +31,4 @@ export class ReceiptService {
     return this.http.put<Receipt>(url + receipt.id, receipt);
   }
 
-  upLoadImage(file: string) {
-    return this.http.post(environment.ApiEndPoint + '/files', '\"' + file + '\"');
-  }
-  getFileById(id: number): Observable<string> {
-    return this.http.get(environment.ApiEndPoint + '/files/' + id);
-  }
-  deleteFileById(id: number) {
-    return this.http.delete(environment.ApiEndPoint + '/files/' + id);
-  }
 }
